Make ublksrv{print|parse}_std_opts public
These two functions are useful for all targets so move them over to
the public API.
