+++ targets/include/ublksrv_tgt.h
@@ -14,7 +14,6 @@
 #include <sys/stat.h>
 #include <sys/ioctl.h>
 
-#include <libgen.h>
 #include <coroutine>
 #include <iostream>
 #include <type_traits>
@@ -178,7 +177,6 @@ struct ublksrv_queue_info {
 	pthread_t thread;
 };
 
-int ublksrv_parse_std_opts(struct ublksrv_dev_data *data, int *efd, int argc, char *argv[]);
 void ublksrv_print_std_opts(void);
 int ublksrv_cmd_dev_add(const struct ublksrv_tgt_type *tgt_type, int argc, char *argv[]);
 char *ublksrv_pop_cmd(int *argc, char *argv[]);
